Add reindexer
