+++ app/(main)/bytes/api.ts
@@ -1,13 +1,13 @@
-// import rehypeShiki from "@stefanprobst/rehype-shiki"
+import rehypeShiki from "@stefanprobst/rehype-shiki"
+import { notFound } from "next/navigation"
 import { compileMDX } from "next-mdx-remote/rsc"
-// import path from "node:path"
+import path from "node:path"
 import { cache } from "react"
 import rehypeSlug from "rehype-slug"
 import remarkGfm from "remark-gfm"
 import remarkSmartypants from "remark-smartypants"
-import { getByteSource } from "lib/api/bytes"
 import prisma from "lib/prisma"
-// import { getHighlighter } from "../../../config/highlighter.mjs"
+import { getHighlighter } from "../../../config/highlighter.mjs"
 import rehypeCodeA11y from "../../../config/rehype-code-a11y.mjs"
 import rehypeCodeTitles from "../../../config/rehype-code-titles.mjs"
 import rehypeHeaderId from "../../../config/rehype-header-id.mjs"
@@ -19,10 +19,15 @@ import { ByteMeta } from "./types"
 export const revalidate = 3600
 
 export const getByte = cache(async (slug: string) => {
-  // const themePath = path.join(process.cwd(), "config/tokyonight.json")
-  // const highlighter = await getHighlighter(themePath)
+  const byte = await prisma.byte.findUnique({ where: { slug } })
+  if (!byte) {
+    notFound()
+  }
 
-  const source = await getByteSource(slug)
+  // Create a shiki highlighter
+  const themePath = path.join(process.cwd(), "config/tokyonight.json")
+  const highlighter = await getHighlighter(themePath)
+
   const { content, frontmatter } = await compileMDX<ByteMeta>({
     options: {
       mdxOptions: {
@@ -31,14 +36,14 @@ export const getByte = cache(async (slug: string) => {
           rehypeHeadings,
           rehypeHeaderId,
           rehypeCodeTitles,
-          // [rehypeShiki, { highlighter }],
+          [rehypeShiki, { highlighter }],
           rehypeCodeA11y,
         ],
         remarkPlugins: [remarkGfm, remarkSmartypants, remarkCodeTitles],
       },
       parseFrontmatter: true,
     },
-    source,
+    source: byte.content,
   })
 
   return { content, meta: frontmatter }
