+++ app/lib/api/github.ts
@@ -12,3 +12,21 @@ octokit.hook.error("request", async (error) => {
 
   throw error
 })
+
+export async function getByteSource(slug: string) {
+  const path = `bytes/${slug}.md`
+  const { data } = await octokit.repos.getContent({
+    mediaType: { format: "raw" },
+    owner: "mskelton",
+    path,
+    repo: "bytes",
+  })
+
+  if (typeof data !== "string") {
+    throw new Error(
+      `Tried to fetch raw file from ${path}. GitHub did not return the raw contents. This should never happen...`,
+    )
+  }
+
+  return data as string
+}
